default number type to double when format is not set

+++ src/Shiny.Mediator.SourceGenerators/Http/OpenApiContractGenerator.cs
@@ -234,7 +234,7 @@ public class OpenApiContractGenerator(MediatorHttpItemConfig itemConfig, Action<
         "int64" => "long",
         "float" => "float",
         "double" => "double",
-        null => "int",
+        null => "double",
         _ => throw new InvalidOperationException("Invalid Number Format - " + format)
     };
